Consolidation of horizon task into multimodal_mimic4.py, removed docs/ entry
